Refactor excel utils.
